+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Ease Animation curves preset
 - Animation Curve Extensions
 - Position (Shake) Animation component
+- Position (Punch) Animation component
 - Position Animation component
 - Scale Animation component
 - TextMesh Pro Visible Characters Animation component